+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+### Added
+
+- Prompt template for transcribing the text in an image. The GPT-4.5-preview model. XML export option.
+
 
 
 ## [1.1.4] – 2025-02-18